requirements.txt
Django==6.0.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
